Big Commit
First new AI autogun, creates a profile of the
AI's unit, as well as scores targets.

New AI follow command based on the A* system.

New chasetarget based on the new AI follow. Also
chasetarget will go into sleep mode when other
AI movement commands are being used, then will
wakeup when that command is done.

Fixed the AI auto sensor system to work with the
new code.

Added a new command 'EVENTSTATS' to AI's interface
allowing a user to debug the various AI events
running on that AI.

Added a #define DEBUG_ASTAR which turns on A*'s
debugging.

Cleaned up muxevent_names which was very out of
date and moved it to mech.events.h

Changed the way physical attacks are handled,
they are more in line with FASA rules, taking
into account torso twist for upper body attacks
as well as allowing punch/sword/axe attacks to
not only occur in the foward arc but their
respective side arc.


git-svn-id: svn://svn.code.sf.net/p/btonline-btech/code@155 e60b412a-5d77-420b-80c2-6cef64185626