Added digittime()
I used MUX's way of format'ing time (look up time_format_1 in 
netcommon.c) as the way to do it.

We can change if this isn't working as expected. IN KM's example, 
dittime(123456) with this comes up as 1d 10:17 (it discards the 36 
seconds instead of the rounding <sort of> that mux2 does).

This is exactly the way the WHO does the time.



git-svn-id: svn://svn.code.sf.net/p/btonline-btech/code@453 e60b412a-5d77-420b-80c2-6cef64185626