Stagger should be less flakey now
btech_newstaggertime for how often we check stagger level

btech_newstaggertons to turn on/off tonnage modifier for stagger

DB Rebuild required.

Code written and submitted by John Fothan

Reworked stagger. Every time you take damage, it is recorded in a linked list stored on the mech object.
Every heartbeat, the damage is checked for validity (is it > 1 turn old? If so, remove it and give me back my memory damnit!). 
Also, during the heartbeat, if X seconds have elapsed (configurable in .conf: btch_newstaggertime) we check the mech for stagger damage.
If the damage they've taken in the last turn is > 20 (or a multiple of it) we clear the smallest amount of damage over a stagger level(multiple of 20. EG: I've taken 8+8+5+3 damage in the last turn. 8+8+5(21) will clear, leaving me with 3 leftover, not 4. 
Had I taken 8+8+3+5, all the damage would clear, however, since the 8+8+3 = 19, not 20, and the whole 5 is required to stagger me).

There is also the ability to scale stagger rolls based on tonnage. <= 35 tons = +1 to stagger BTH. 35 < tons <= 55 = normal stagger bth. 55 < tons <= 75 = -1 to stagger bth. 75 < tons = -2 to stagger bth. (btech_newstaggertons).


git-svn-id: svn://svn.code.sf.net/p/btonline-btech/code@1326 e60b412a-5d77-420b-80c2-6cef64185626