Several changes that will require a db fix/hardshutdown. (Can all be avoided if you save data in your mux like damages, stores, and mechtype and reload them after):
Addition of Light_Bap sensor (Sensor is 'a' instead of 'b'). Range 3, same as bap for the rest of the rules
HSEngOverRide xcodevalue, for rare (and i do mean rare) mechs that don't use the default engine Heatsinks (Ryoken-Prime is a good example)
Change in mech ref length (adding 10 for long clan names)
Removal of 'Loader_Tech'. Isn't using anywhere in the hardcode.


git-svn-id: svn://svn.code.sf.net/p/btonline-btech/code@1409 e60b412a-5d77-420b-80c2-6cef64185626